todoのタイルをコンポーネント化した
