+++ lib/provider/todos.dart
@@ -1,5 +1,4 @@
 import 'package:kagb_template/model/todo.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'todos.g.dart';
