+++ lib/view/home_screen.dart
@@ -1,0 +1,82 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:kagb_template/model/todo.dart';
+import 'package:kagb_template/provider/todos.dart';
+import 'package:kagb_template/view/components/todo_card.dart';
+import 'package:uuid/uuid.dart';
+
+class HomeScreen extends ConsumerWidget {
+  const HomeScreen({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final TextEditingController todoController = TextEditingController();
+
+    return SafeArea(
+      child: Scaffold(
+        appBar: AppBar(
+          title: const Text('Todo App'),
+          backgroundColor: Colors.blue,
+        ),
+        body: Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Center(
+            child: Column(
+              children: [
+                Padding(
+                  padding: const EdgeInsets.symmetric(vertical: 15.0),
+                  child: Consumer(
+                    builder: (context, ref, child) {
+                      return TextField(
+                        controller: todoController,
+                        decoration: const InputDecoration(
+                          labelText: 'Enter your Todo',
+                          border: OutlineInputBorder(),
+                        ),
+                        onSubmitted: (text) => submitTodo(text, ref),
+                      );
+                    },
+                  ),
+                ),
+                Expanded(
+                  child: Consumer(
+                    builder: (context, ref, child) {
+                      final todos = ref.watch(todosProvider);
+                      return ListView.builder(
+                        itemBuilder: (BuildContext context, int index) {
+                          return TodoCard(
+                            todo: todos[index],
+                            deleteTodo: (id) => deleteTodo(ref, id),
+                          );
+                        },
+                        itemCount: todos.length,
+                      );
+                    },
+                  ),
+                )
+              ],
+            ),
+          ),
+        ),
+        floatingActionButton: FloatingActionButton(
+          onPressed: () => print("押された"),
+          child: const Icon(Icons.add),
+        ),
+      ),
+    );
+  }
+
+  submitTodo(String todo, WidgetRef ref) {
+    // TODO: 入力されたTODOをViewModelに通知し、登録する
+    final newTodo = Todo(
+      todo: todo,
+      id: const Uuid().v4(),
+      isCompleted: false,
+    );
+    ref.read(todosProvider.notifier).addTodo(newTodo);
+  }
+
+  deleteTodo(WidgetRef ref, String id) {
+    ref.read(todosProvider.notifier).deleteTodoById(id);
+  }
+}
